KPI Monitor fixed and working, e2 simulator improved

+++ RAN_Intelligent_Controllers/Near-Real-Time-RIC/install_scripts/run_e2sim_and_connect_to_ric.sh
@@ -40,21 +40,30 @@ cd "$(dirname "$SCRIPT_DIR")"
 mkdir -p logs
 OUTPUT_FILE="logs/e2sim_output.txt"
 
+# Optionally, stop the oransim container before starting it again
+# if [ $(sudo docker ps -q -f name=^/oransim$ | wc -l) -eq 1 ]; then
+#     echo "Restarting oransim container..."
+#     sudo docker stop oransim
+# fi
+
 # Check if the container with the name 'oransim' is already running
 if [ $(sudo docker ps -q -f name=^/oransim$ | wc -l) -eq 1 ]; then
     echo "Container 'oransim' is already running."
 elif [ $(sudo docker ps -aq -f name=^/oransim$ | wc -l) -eq 1 ]; then
     echo "Container 'oransim' exists but is not running, starting container..."
-    rm -rf OUTPUT_FILE
+    rm -rf $OUTPUT_FILE
     sudo docker start oransim
 else
     echo "Starting a new container 'oransim'..."
-    rm -rf OUTPUT_FILE
+    rm -rf $OUTPUT_FILE
     sudo docker run -d -it --name oransim oransim:0.0.999
 fi
 kubectl get svc -n ricplt | grep e2term-sctp || true
 
-sudo docker exec oransim pkill -f kpm_sim || true
+if sudo docker exec oransim pgrep -f "kpm_sim" >/dev/null; then
+    echo "Stopping previous instance of kpm_sim..."
+    sudo docker exec oransim pkill -f kpm_sim || true
+fi
 
 # Get the IP and port of the E2 termination point inside the near Real Time RIC
 SERVICE_NAME="service-ricplt-e2term-sctp"
@@ -64,12 +73,11 @@ ATTEMPTS=1
 MAX_ATTEMPTS=10
 KPM_RESTARTS=1
 KPM_MAX_RESTARTS=3
-export CHART_REPO_URL=http://0.0.0.0:8090
 
 # Monitor output file for a success message
 while true; do
     if [ -z "$LINE" ]; then
-        LINE=$(kubectl get svc -n ricplt | grep $SERVICE_NAME) || ""
+        LINE=$(kubectl get svc -n ricplt | grep $SERVICE_NAME)
         IP_E2TERM=$(echo $LINE | awk '{print $3}')
         PORT_E2TERM=$(echo $LINE | awk '{print $5}' | sed 's/:.*//')
         echo "IP for $SERVICE_NAME: $IP_E2TERM"
@@ -88,10 +96,10 @@ while true; do
         sudo chown $USER:$USER $OUTPUT_FILE
     fi
     # Check if kpm_sim is already running to avoid duplicate runs
-    if ! pgrep -f "kpm_sim $IP_E2TERM $PORT_E2TERM" >/dev/null; then
+    if ! sudo docker exec oransim pgrep -f "kpm_sim" >/dev/null; then
         echo "Starting kpm_sim in the background, writing to $OUTPUT_FILE..."
         >"$OUTPUT_FILE" # Clears the content of the output file
-        sudo docker exec -i oransim kpm_sim $IP_E2TERM $PORT_E2TERM >$OUTPUT_FILE 2>&1 &
+        sudo docker exec -i oransim kpm_sim $IP_E2TERM $PORT_E2TERM >>$OUTPUT_FILE 2>&1 &
         sleep 2
     fi
 
@@ -106,9 +114,11 @@ while true; do
     if [ "$ATTEMPTS" -eq "$MAX_ATTEMPTS" ]; then
         cat $OUTPUT_FILE
         kubectl get pods -A || true
-        echo
-        echo "Restarting kpm_sim inside of oransim..."
-        sudo docker exec oransim pkill -f kpm_sim || true
+        if sudo docker exec oransim pgrep -f "kpm_sim" >/dev/null; then
+            echo
+            echo "Restarting kpm_sim inside of oransim..."
+            sudo docker exec oransim pkill -f kpm_sim || true
+        fi
         ATTEMPTS=0
         KPM_RESTARTS=$((KPM_RESTARTS + 1))
         if [ "$KPM_RESTARTS" -eq "$KPM_MAX_RESTARTS" ]; then
@@ -121,8 +131,11 @@ while true; do
     ATTEMPTS=$((ATTEMPTS + 1))
 done
 
+# If console breaks, reset it
+stty sane || true
+
 SERVICE_NAME="service-ricplt-e2mgr-http"
-LINE=$(kubectl get svc -n ricplt | grep $SERVICE_NAME) || ""
+LINE=$(kubectl get svc -n ricplt | grep $SERVICE_NAME || echo "")
 IP_HTTP_E2TERM=$(echo $LINE | awk '{print $3}')
 PORT_HTTP_E2TERM=$(echo $LINE | awk '{print $5}' | sed 's/\/.*//' | cut -d: -f2)
 echo "$IP_HTTP_E2TERM:$PORT_HTTP_E2TERM"
